flightsite/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mysqlclient==2.1.1
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pytz==2022.7.1
rfc3986==1.5.0
rich==12.6.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.7
